src/main/java/com/exasol/adapter/dialects/snowflake/SnowflakeSqlDialectFactory.java
package com.exasol.adapter.dialects.snowflake;

import com.exasol.adapter.AdapterProperties;
import com.exasol.adapter.dialects.AbstractSqlDialectFactory;
import com.exasol.adapter.dialects.SqlDialect;
import com.exasol.adapter.jdbc.ConnectionFactory;

public class SnowflakeSqlDialectFactory extends AbstractSqlDialectFactory {
    @Override
    public String getSqlDialectName() {
        return SnowflakeSqlDialect.NAME;
    }

    @Override
    public SqlDialect createSqlDialect(final ConnectionFactory connectionFactory, final AdapterProperties properties) {
        return new SnowflakeSqlDialect(connectionFactory, properties);
    }
}